requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
